requirements.txt
snowflake-connector-python==3.15
openai==1.97.1
pandas==2.2.2
python-dotenv==1.0.1